pyproject.toml
[tool.poetry]
name = "stratux-companion"
version = "0.1.0"
description = ""
authors = ["Alexander Gorokhov"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
websockets = "^12.0"
luma-lcd = "^2.11.0"
pydantic = "<2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
